Complete CRUD for blogs, add login, add firebaseUploadService

+++ Application/Handlers/Tags/DeleteTagHandler.cs
@@ -1,15 +1,10 @@
 ﻿using Application.Commands.Tags;
-using Application.Mappers;
 using Application.Response;
 using Application.Response.Base;
-using Core.Entities;
 using Core.Repositories;
 using MediatR;
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -29,27 +24,33 @@ namespace Application.Handlers.Tags
             var response = new Response<TagResponse>();
             try
             {
-                var result = await _tagRepository.GetByIdAsync(request.TagId);
-
+                var result = await _tagRepository.GetByIdAsync(request.Id);
+                if (result is null)
+                {
+                    throw new NullReferenceException("Not found tag");
+                }
                 await _tagRepository.DeleteAsync(result);
-                response = new Response<TagResponse>() {
-                    StatusCode = HttpStatusCode.OK
-                };
 
+                response = new Response<TagResponse>()
+                {
+                    StatusCode = HttpStatusCode.NoContent
+                };
             }
-            catch (ApplicationException ex)
+            catch (NullReferenceException ex)
             {
-                response = new Response<TagResponse>(ex.Message);
-                response.StatusCode = HttpStatusCode.UnprocessableEntity;
+                response = new Response<TagResponse>(ex.Message)
+                {
+                    StatusCode = HttpStatusCode.NotFound
+                };
             }
             catch (Exception ex)
             {
-                response = new Response<TagResponse>(ex.Message);
-                response.StatusCode = HttpStatusCode.InternalServerError;
+                response = new Response<TagResponse>(ex.Message)
+                {
+                    StatusCode = HttpStatusCode.InternalServerError
+                };
             }
-
             return response;
-
         }
     }
 }
